refactor: viewmodel usage
